NTBS-2246: Split steps into separate classes and add imported table name into config
