+++ ntbs-service/DataAccess/AlertRepository.cs
@@ -72,8 +72,7 @@ namespace ntbs_service.DataAccess
         {
             return await GetBaseOpenAlertIQueryable()
                 .OfType<TransferAlert>()
-                .Include(a => a.TbService)
-                .ThenInclude(tbs => tbs.PHEC)
+                .Include(a => a.TbService.PHEC)
                 .Include(a => a.CaseManager)
                 .Where(a => a.NotificationId == notificationId)
                 .SingleOrDefaultAsync();
