+++ ntbs-ui-tests/Hooks/DriverSetup.cs
@@ -29,7 +29,7 @@ namespace ntbs_ui_tests.Hooks
         }
 
         [BeforeScenario]
-        public async void BeforeScenario()
+        public void BeforeScenario()
         {
             var opts = new ChromeOptions();
             opts.AddArgument("--no-sandbox"); // Necessary to avoid `unknown error: DevToolsActivePort file doesn't exist` when running on docker
@@ -41,7 +41,6 @@ namespace ntbs_ui_tests.Hooks
             Browser = new RemoteWebDriver(opts);
             Browser.Manage().Timeouts().ImplicitWait = settings.ImplicitWait;
             objectContainer.RegisterInstanceAs(Browser);
-            await CleanUpMigratedNotification();
         }
 
         [AfterScenario]
@@ -56,6 +55,19 @@ namespace ntbs_ui_tests.Hooks
                 await connection.ExecuteAsync(deleteNotifications, new { ids = testContext.AddedNotificationIds.ToArray() });
             }
             Browser.Quit();
+        }
+
+        [Scope(Feature = "Import legacy notification")]
+        [BeforeScenario]
+        private async Task RemoveImportedNotificationBefore()
+        {
+            await CleanUpMigratedNotification();
+        }
+
+        [Scope(Feature = "Import legacy notification")]
+        [AfterScenario]
+        private async Task RemoveImportedNotificationAfter()
+        {
             await CleanUpMigratedNotification();
         }
 
@@ -70,19 +82,10 @@ namespace ntbs_ui_tests.Hooks
             }
             using (var connection = new SqlConnection(settings.EnvironmentConfig.MigrationConnectionString))
             {
-                var importedNotificationTableName = GetImportedNotificationTableName();
                 connection.Open();
-                var deleteImportedNotification = $"DELETE FROM {importedNotificationTableName} WHERE LegacyId = '189045'";
+                var deleteImportedNotification = $"DELETE FROM {settings.EnvironmentConfig.ImportedNotificationTableName} WHERE LegacyId = '189045'";
                 await connection.ExecuteAsync(deleteImportedNotification);
             }
-        }
-
-        private string GetImportedNotificationTableName()
-        {
-            var importedTablePrefix = settings.EnvironmentUnderTest == "local"
-                ? "Dev"
-                : char.ToUpper(settings.EnvironmentUnderTest[0]) + settings.EnvironmentUnderTest.Substring(1);
-            return $"{importedTablePrefix}ImportedNotifications";
         }
     }
 }
