+++ ntbs-ui-tests/testSettings.json
@@ -13,7 +13,8 @@
     "local": {
       "RootUri": "https://localhost:5001",
       "ConnectionString": "data source=localhost;initial catalog=ntbsDev;trusted_connection=true;MultipleActiveResultSets=true;",
-      "MigrationConnectionString": "data source=localhost;initial catalog=ntbs_migration;trusted_connection=true;MultipleActiveResultSets=true"
+      "MigrationConnectionString": "data source=localhost;initial catalog=ntbs_migration;trusted_connection=true;MultipleActiveResultSets=true",
+      "ImportedNotificationTableName": "DevImportedNotifications"
     },
     "int": {
       "RootUri": "https://ntbs-int.e32846b1ddf0432eb63f.northeurope.aksapp.io"
